fix(UI): fix moment.js date formatting for "Poikkileikkauspäivä"
The Linjakarttageneraattori poikkileikkaupaiva calendar picker tool
displayed wrongly formatted date in UI and the same date got sent to
back-end as well. This caused the functionality of the date picker to break.

This commit changes the way how moment.js formats the day in PointBuildTrigger.js.
Date formatting with moment.js is already done in DayPicker component and "reformatting"
it twice in PointBuildTrigger caused the bug.